extras/plugins/nancy-provider-codex/pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "nancy-provider-codex"
version = "0.1.0"
description = "Native OpenAI Codex (backend-api) provider for Nexus-Nancy"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "AmberLee2427" }
]
dependencies = [
    "nexus-nancy>=1.1.1",
    "httpx>=0.27.0",
    "PyJWT>=2.8.0"
]

[project.entry-points."nexus_nancy.providers"]
codex = "nancy_codex.plugin"

[project.entry-points."nexus_nancy.plugins"]
codex_auth = "nancy_codex.plugin"

[tool.hatch.build.targets.wheel]
packages = ["src/nancy_codex"]
